Create Method_Syntax
Methods are functions that are defined within the context of a struct. The first parameter is always self, which represents the instance of the struct the method is being called on.

A great technique in organizing the capabilities in what can be done in an Instance of a type in one impl block.