Update Static_Lifetime
A static variable is a memory resource created at compile-time that exists through a program start to finish. They must have their types explicitly specified.

A static lifetime is a memory resource that lasts indefinitely to the end of a program. Note that by this definition some static lifetime resources can be created at runtime.