Update Packages_Crates
Module
A module is a language construct that acts as the basic building block of code organization within a crate. A module is a container for functions, types, and nested modules. Modules also specify the visibility for all the items that they contain or re-export.
Crate
A crate is the basic unit of compilation and linking. Crates are also part of the language (crate is a keyword), but you usually don't mention them much in your source code. There are two main types of crates: libraries and executable files.
Package
A package is the basic unit of software distribution. Packages are not part of the language, so you will not find them in the language reference. Packages are artifacts of the Rust package manager, cargo. Packages can contain one or more crates: at most one library and any number of executables.